+++ .gitignore
@@ -1,1 +1,2 @@
 search.py
+__pycache__
