+++ main.py
@@ -1,4 +1,5 @@
-from search import github_search
+from search.search import github_search
+
 
 if __name__ == "__main__":
     github_search()
