+++ search/search_by_repository.py
@@ -1,11 +1,11 @@
 import requests
 from log import *
-from repository_clone import github_crawler
-from search_option import *
+from search.repository_clone import *
+from search.search_option import *
 
 
 def search_by_repository(headers):
-    from search import log_txt
+    from search.search import log_txt
 
     save_repository_info("Search type : Repository\n", log_txt)
     option = search_option()
